backend/requirements.txt
fastapi
uvicorn[standard]
azure-ai-textanalytics
azure-core
azure-data-tables
WebSocket
gunicorn
datetime